% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.adaptive_masses}
\alias{.adaptive_masses}
\title{Returns the sizes of the adaptive bins used for the adaptive ECBC for one vector.}
\usage{
.adaptive_masses(X, resolution)
}
\arguments{
\item{X}{A vector, representing one sample of one variable}

\item{resolution}{The resolution of the CB approximation}
}
\value{
A numeric vector of bin sizes
}
\description{
Returns the sizes of the adaptive bins used for the adaptive ECBC for one vector.
}
