% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.CB_make_cumulative_df}
\alias{.CB_make_cumulative_df}
\title{Returns a list of reverse cumulative margins of a CB copula. The nth entry is thus the copula of X1,...,Xn}
\usage{
.CB_make_cumulative_df(CB)
}
\arguments{
\item{CB}{A matrix of CB weights.}
}
\value{
A list of CB weight matrixes of ascending dimension
}
\description{
Returns a list of reverse cumulative margins of a CB copula. The nth entry is thus the copula of X1,...,Xn
}
