% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.EACBC}
\alias{.EACBC}
\title{Calculates an empirical CB approximation with adaptive bin sizes. This will be faster on data with many ties.}
\usage{
.EACBC(X, resolution)
}
\arguments{
\item{X}{A nxrho matrix of n samples of rho variables}

\item{resolution}{The resolution of the CB approximation}
}
\value{
A matrix of dimension resolution^rho
}
\description{
Calculates an empirical CB approximation with adaptive bin sizes. This will be faster on data with many ties.
}
