% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.EACBC_nonzero}
\alias{.EACBC_nonzero}
\title{Returns non 0 entries of the EACBC}
\usage{
.EACBC_nonzero(X, resolution)
}
\arguments{
\item{X}{A nxrho matrix of n samples of rho variables}

\item{resolution}{The resolution of the CB approximation}
}
\value{
A list of local kernel masses
}
\description{
Returns non 0 entries of the EACBC
}
