% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ECBC}
\alias{.ECBC}
\title{Calculates the empirical checkerboard approximation to some data.}
\usage{
.ECBC(X, resolution)
}
\arguments{
\item{X}{A nxrho matrix of n samples of rho variables}

\item{resolution}{The resolution of the CB approximation}
}
\value{
A matrix of dimension resolution^rho
}
\description{
Calculates the empirical checkerboard approximation to some data.
}
