% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.local_kernel_integral}
\alias{.local_kernel_integral}
\title{Computes the D1-difference of two CB matrizes on a local CB dimension}
\usage{
.local_kernel_integral(k1, k2, y)
}
\arguments{
\item{k1}{Vector of local CB weights of first matrix}

\item{k2}{Vector of local CB weights of second matrix}

\item{y}{Vector indicating the bin sizes of the local dimension}
}
\value{
number indicating the difference between k1 and k2
}
\description{
Computes the D1-difference of two CB matrizes on a local CB dimension
}
