% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.random_CB}
\alias{.random_CB}
\title{Creates a random CB copula of resolution 2^steps}
\usage{
.random_CB(rho, steps, de, ie)
}
\arguments{
\item{rho}{The number of variables}

\item{steps}{Number of iteration steps, the final resolution will be 2^steps}

\item{de}{Exponent to increase dependence}

\item{ie}{Exponent to increase independence}
}
\value{
A matrix of dimension (2^steps)^rho
}
\description{
Creates a random CB copula of resolution 2^steps
}
