% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.sample_CB}
\alias{.sample_CB}
\title{Generate a sample of some CB copula-}
\usage{
.sample_CB(CB, n)
}
\arguments{
\item{CB}{A weight matrix of a CB copula}

\item{n}{The number of samples to be generated}
}
\value{
Matrix of dimension nxm where m is the dimension of CB
}
\description{
Generate a sample of some CB copula-
}
