% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qmdrank}
\alias{qmdrank}
\title{Equivalent to rank(x, ties.method="max"), but considerably faster}
\usage{
qmdrank(x)
}
\arguments{
\item{x}{Numeric Vector}
}
\value{
A vector of same length of x of elements 1:length(x)
}
\description{
Equivalent to rank(x, ties.method="max"), but considerably faster
}
