% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_until_changes}
\alias{seq_until_changes}
\title{Returns a vector}
\usage{
seq_until_changes(x)
}
\arguments{
\item{x}{A usually sorted vector}
}
\value{
A sequence along x. If consecutive values in x are equal the maximal value is used.
}
\description{
Returns a vector
}
