% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{qmdrank}
\alias{qmdrank}
\title{Equivalent to rank(x, ties.method = "max") but not as stupidly slow}
\usage{
qmdrank(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
An integer vector specifying for each value in x the rank
        within x. If one value appears multiple time the maximum
        is used.
}
\description{
Equivalent to rank(x, ties.method = "max") but not as stupidly slow
}
